+++ src/main/java/com/clubber/ClubberServer/global/jwt/JwtTokenUtil.java
@@ -4,7 +4,7 @@ import com.clubber.ClubberServer.domain.user.domain.AccountRole;
 import com.clubber.ClubberServer.domain.user.exception.InvalidTokenException;
 import com.clubber.ClubberServer.domain.user.exception.RefreshTokenExpiredException;
 import com.clubber.ClubberServer.domain.user.exception.TokenExpiredException;
-import com.clubber.ClubberServer.global.dto.AccessTokenInfo;
+import com.clubber.ClubberServer.global.jwt.vo.AccessTokenInfo;
 import com.clubber.ClubberServer.global.properties.JwtProperties;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.ExpiredJwtException;
